backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0
pymongo==4.6.1
python-dotenv==1.0.0
apscheduler==3.10.4
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.4
twilio==8.10.0
requests==2.31.0
pytz==2023.3